# spec/microservices/ms-02/04-api-contract.yaml
openapi: 3.1.0
info:
  title: Catalog and Product API
  version: 1.0.0
  description: |
    API contract for MS-02 Catalog and Product.
    This is the single source of truth for paths, fields, status codes, and response shapes.
servers:
  - url: http://localhost:8102/api/v1
...
paths:
...
  /products/{productId}/availability:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
    get:
      operationId: getProductAvailability
      summary: Get regional product availability.
      security: []
      responses:
        '200':
          description: Availability list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: replaceProductAvailability
      summary: Replace product availability.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceAvailabilityRequest' }
      responses:
        '200':
          description: Availability list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /products/{productId}/reservations:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
    post:
      operationId: createInventoryReservation
      summary: Atomically reserve product or variant availability.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReservationRequest' }
      responses:
        '201':
          description: Held reservation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryReservation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /reservations/{reservationId}/commit:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ReservationId'
    post:
      operationId: commitInventoryReservation
      summary: Commit a held inventory reservation.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Committed reservation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryReservation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /reservations/{reservationId}/release:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ReservationId'
    post:
      operationId: releaseInventoryReservation
      summary: Release a held inventory reservation.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Released reservation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryReservation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /products/{productId}/media:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
    post:
      operationId: uploadProductMedia
      summary: Upload product media through the configured provider.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ProductMediaUploadRequest' }
          application/json:
            schema: { $ref: '#/components/schemas/ExternalMediaRequest' }
      responses:
        '201':
          description: Product media.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductMedia' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /products/{productId}/media/{mediaId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/ProductId'
      - $ref: '#/components/parameters/MediaId'
    delete:
      operationId: deleteProductMedia
      summary: Delete product media and publish a projection change.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Deletion result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletionResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /categories:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
    get:
      operationId: listCategories
      summary: List categories and hierarchy.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/LanguageCode'
        - name: name
          in: query
          schema: { type: string }
        - name: visible
          in: query
          schema: { type: boolean }
        - name: featured
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Category page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryListResponse' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    post:
      operationId: createCategory
      summary: Create a category.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryRequest' }
      responses:
        '201':
          description: Category created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /categories/{categoryId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/CategoryId'
    get:
      operationId: getCategory
      summary: Get a category and its children.
      security: []
      parameters:
        - $ref: '#/components/parameters/LanguageCode'
      responses:
        '200':
          description: Category.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateCategory
      summary: Update category metadata.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCategoryRequest' }
      responses:
        '200':
          description: Updated category.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      operationId: deleteCategory
      summary: Delete a category subtree.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orphanProductPolicy
          in: query
          required: false
          schema:
            type: string
            enum: [Detach, Delete, Reject]
            default: Reject
      responses:
        '200':
          description: Category deletion result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryDeletionResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /categories/slug/{friendlyUrl}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/FriendlyUrl'
    get:
      operationId: getCategoryByFriendlyUrl
      summary: Get a category by friendly URL.
      security: []
      parameters:
        - $ref: '#/components/parameters/LanguageCode'
      responses:
        '200':
          description: Category.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404': { $ref: '#/components/responses/NotFound' }
  /categories/uniqueness:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
    get:
      operationId: checkCategoryCodeUniqueness
      summary: Check category code uniqueness.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 100 }
        - name: excludeCategoryId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Existence result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExistsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /categories/{categoryId}/visibility:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/CategoryId'
    patch:
      operationId: updateCategoryVisibility
      summary: Update category visibility.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [visible]
              properties:
                visible: { type: boolean }
      responses:
        '200':
          description: Updated category.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /categories/{categoryId}/move/{parentId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/CategoryId'
      - $ref: '#/components/parameters/ParentId'
    put:
      operationId: moveCategory
      summary: Move a category and recalculate descendants.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Moved category.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /categories/{categoryId}/products:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/StoreId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/CategoryId'
    get:
      operationId: listCategoryProducts
      summary: List products in a category subtree.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/LanguageCode'
        - $ref: '#/components/parameters/CountryCode'
      responses:
        '200':
          description: Product page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      $ref: '../../shared/common-schemas.yaml#/components/securitySchemes/bearerAuth'
  parameters:
    TenantId:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/TenantId'
    StoreId:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/StoreId'
    CorrelationId:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/CorrelationId'
    IdempotencyKey:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/IdempotencyKey'
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ParentId:
      name: parentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ReservationId:
      name: reservationId
      in: path
      required: true
      schema: { type: string, format: uuid }
    MediaId:
      name: mediaId
      in: path
      required: true
      schema: { type: string, format: uuid }
    FriendlyUrl:
      name: friendlyUrl
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    LanguageCode:
      name: languageCode
      in: query
      required: false
      schema: { type: string, default: en }
    CountryCode:
      name: countryCode
      in: query
      required: false
      schema: { type: string, minLength: 2, maxLength: 2 }
    Page:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/Page'
    PageSize:
      $ref: '../../shared/common-schemas.yaml#/components/parameters/PageSize'
  responses:
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient catalog permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found or not eligible in the requested scope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Business or concurrency conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableEntity:
      description: Validation failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ServiceUnavailable:
      description: Media or dependent provider unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      $ref: '../../shared/common-schemas.yaml#/components/schemas/ErrorResponse'
    PaginationInfo:
      $ref: '../../shared/common-schemas.yaml#/components/schemas/PaginationInfo'
    ProductListResponse:
      allOf:
        - $ref: '../../shared/common-schemas.yaml#/components/schemas/ListEnvelope'
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Product' }
        pagination: { $ref: '#/components/schemas/PaginationInfo' }
    CategoryListResponse:
      allOf:
        - $ref: '../../shared/common-schemas.yaml#/components/schemas/ListEnvelope'
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Category' }
        pagination: { $ref: '#/components/schemas/PaginationInfo' }
...
    ExistsResponse:
      type: object
      required: [exists]
      properties:
        exists: { type: boolean }
    DeletionResult:
      type: object
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [Deleted] }
        dependentsRemoved: { type: integer, minimum: 0 }
        projectionEventPublished: { type: boolean }
    CategoryDeletionResult:
      type: object
      required: [categoryId, status, deletedCategoryCount]
      properties:
        categoryId: { type: string, format: uuid }
        status: { type: string, enum: [Deleted] }
        deletedCategoryCount: { type: integer, minimum: 1 }
        detachedProductCount: { type: integer, minimum: 0 }
        deletedProductCount: { type: integer, minimum: 0 }
    ProductDescription:
      type: object
      required: [languageCode, name, friendlyUrl]
      properties:
        languageCode: { type: string, minLength: 2 }
        name: { type: string, minLength: 1 }
        friendlyUrl: { type: string, minLength: 1 }
        description: { type: string }
        highlights: { type: string }
        title: { type: string }
        keywords: { type: string }
        metaDescription: { type: string }
    ProductMedia:
      type: object
      required: [id, fileName, imageType, mediaStatus]
      properties:
        id: { type: string, format: uuid }
        fileName: { type: string }
        imageType: { type: string, enum: [Binary, ExternalUrl] }
        originalUri: { type: string, format: uri }
        transformedUri: { type: string, format: uri }
        externalUrl: { type: string, format: uri }
        defaultImage: { type: boolean }
        mediaStatus: { type: string, enum: [Pending, Ready, Failed, Deleted] }
    Availability:
      type: object
      required: [id, regionCode, quantity, reservedQuantity, active]
      properties:
        id: { type: string, format: uuid }
        regionCode: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 0 }
        reservedQuantity: { type: integer, minimum: 0 }
        sellableQuantity: { type: integer, minimum: 0 }
        active: { type: boolean }
    AvailabilityListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Availability' }
...
    ProductOptionValue:
      type: object
      required: [id, code, displayOnly]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        displayOnly: { type: boolean }
        priceAdjustment: { type: number }
        imageUri: { type: string, format: uri }
    ProductOption:
      type: object
      required: [id, code, displayOnly, values]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        displayOnly: { type: boolean }
        values:
          type: array
          items: { $ref: '#/components/schemas/ProductOptionValue' }
    ProductVariant:
      type: object
      required: [id, productId, sku, status, defaultSelection]
      properties:
        id: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        sku: { type: string }
        code: { type: string }
        status: { type: string, enum: [Draft, Active, Unavailable, Deleted] }
        defaultSelection: { type: boolean }
        available: { type: boolean }
        dateAvailable: { type: string, format: date-time }
        availability:
          type: array
          items: { $ref: '#/components/schemas/Availability' }
        price: { $ref: '#/components/schemas/Price' }
    Product:
      type: object
      required: [id, storeId, sku, status, visible, available, canBePurchased, dateAvailable, descriptions, availabilities]
      properties:
        id: { type: string, format: uuid }
        storeId: { type: string }
        sku: { type: string }
        refSku: { type: string }
        status: { type: string, enum: [Draft, Active, Unavailable, Deleted] }
        visible: { type: boolean }
        available: { type: boolean }
        canBePurchased: { type: boolean }
        dateAvailable: { type: string, format: date-time }
        manufacturerCode: { type: string }
        productTypeCode: { type: string }
        taxClassCode: { type: string }
        productVirtual: { type: boolean }
        productShippable: { type: boolean }
        productFree: { type: boolean }
        length: { type: number }
        width: { type: number }
        height: { type: number }
        weight: { type: number }
        reviewAverage: { type: number }
        reviewCount: { type: integer, minimum: 0 }
        sortOrder: { type: integer }
        descriptions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ProductDescription' }
        categories:
          type: array
          items: { $ref: '#/components/schemas/CategoryReference' }
        variants:
          type: array
          items: { $ref: '#/components/schemas/ProductVariant' }
        availabilities:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Availability' }
        options:
          type: array
          items: { $ref: '#/components/schemas/ProductOption' }
        properties:
          type: array
          items: { $ref: '#/components/schemas/ProductProperty' }
        media:
          type: array
          items: { $ref: '#/components/schemas/ProductMedia' }
        price: { $ref: '#/components/schemas/Price' }
    ProductProperty:
      type: object
      required: [code, value]
      properties:
        code: { type: string }
        value: { type: string }
        name: { type: string }
    CategoryReference:
      type: object
      required: [id, code]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
    Category:
      type: object
      required: [id, storeId, code, status, visible, depth, lineage, descriptions]
      properties:
        id: { type: string, format: uuid }
        storeId: { type: string }
        code: { type: string }
        parentId: { type: string, format: uuid }
        status: { type: string, enum: [Draft, Active, Hidden, Deleted] }
        visible: { type: boolean }
        featured: { type: boolean }
        sortOrder: { type: integer }
        depth: { type: integer, minimum: 0 }
        lineage: { type: string }
        descriptions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/CategoryDescription' }
        children:
          type: array
          items: { $ref: '#/components/schemas/CategoryReference' }
    CategoryDescription:
      type: object
      required: [languageCode, name, friendlyUrl]
      properties:
        languageCode: { type: string }
        name: { type: string }
        friendlyUrl: { type: string }
        description: { type: string }
        title: { type: string }
        metaDescription: { type: string }
    InventoryReservation:
      type: object
      required: [id, productId, quantity, state, expiresAt]
      properties:
        id: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        variantId: { type: string, format: uuid }
        availabilityId: { type: string, format: uuid }
        reservationKey: { type: string }
        quantity: { type: integer, minimum: 1 }
        state: { type: string, enum: [Held, Committed, Released, Expired] }
        expiresAt: { type: string, format: date-time }
        committedAt: { type: string, format: date-time }
        releasedAt: { type: string, format: date-time }
...
    AvailabilityInput:
      type: object
      required: [regionCode, quantity]
      properties:
        regionCode: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 0 }
        active: { type: boolean, default: true }
    ReplaceAvailabilityRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/AvailabilityInput' }
...
    CreateReservationRequest:
      type: object
      required: [reservationKey, quantity, expiresAt]
      properties:
        reservationKey: { type: string, minLength: 1, maxLength: 200 }
        variantId: { type: string, format: uuid }
        availabilityId: { type: string, format: uuid }
        regionCode: { type: string, default: '*' }
        quantity: { type: integer, minimum: 1 }
        expiresAt: { type: string, format: date-time }
    ProductMediaUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        fileName: { type: string }
        defaultImage: { type: boolean, default: false }
    ExternalMediaRequest:
      type: object
      required: [externalUrl, fileName]
      properties:
        externalUrl: { type: string, format: uri }
        fileName: { type: string }
        defaultImage: { type: boolean, default: false }
    CreateCategoryRequest:
      type: object
      required: [code, descriptions]
      properties:
        code: { type: string, minLength: 1, maxLength: 100, pattern: '^[A-Za-z0-9_-]+$' }
        parentId: { type: string, format: uuid }
        visible: { type: boolean, default: false }
        featured: { type: boolean, default: false }
        sortOrder: { type: integer, default: 0 }
        descriptions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/CategoryDescription' }
    UpdateCategoryRequest:
      type: object
      minProperties: 1
      properties:
        code: { type: string }
        parentId: { type: string, format: uuid }
        visible: { type: boolean }
        featured: { type: boolean }
        sortOrder: { type: integer }
        descriptions:
          type: array
          items: { $ref: '#/components/schemas/CategoryDescription' }
...